[Refactor] 아이젠하워 매트릭스 화면 수정
- 사이드바 고정식으로 수정
- trello 형식으로 아이젠하워 매트릭스를 수정함
- 일정 클릭시 상세정보 페이지로 이동
- 삭제된 일정은 보이지 않도록 수정함
